Refactor Routes & Control Logic
